# generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "fnirs-flow API",
    "description": "fNIRS analysis Flow orchestration framework",
    "version": "1.3.0"
  },
  "paths": {
...
    "/api/projects/{project_id}/results/{kind}": {
      "get": {
        "summary": "Project Results Endpoint",
        "operationId": "project_results_endpoint_api_projects__project_id__results__kind__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kind"
            }
          },
          {
            "name": "row_limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Row Limit"
            }
          },
          {
            "name": "file_limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "File Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResults"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjectResults": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "qc",
              "channel",
              "roi",
              "group"
            ],
            "title": "Kind"
          },
          "file_count": {
            "type": "integer",
            "title": "File Count"
          },
          "returned_file_count": {
            "type": "integer",
            "title": "Returned File Count",
            "default": 0
          },
          "truncated": {
            "type": "boolean",
            "title": "Truncated",
            "default": false
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/ResultFile"
            },
            "type": "array",
            "title": "Files"
          },
          "figures": {
            "items": {
              "$ref": "#/components/schemas/ResultFigure"
            },
            "type": "array",
            "title": "Figures"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "file_count"
        ],
        "title": "ProjectResults"
      },
...
      "ResultFigure": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "svg": {
            "type": "string",
            "title": "Svg"
          }
        },
        "type": "object",
        "required": [
          "path",
          "svg"
        ],
        "title": "ResultFigure"
      },
      "ResultFile": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "data": {
            "title": "Data"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes",
            "default": 0
          },
          "returned_rows": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Returned Rows"
          },
          "rows_truncated": {
            "type": "boolean",
            "title": "Rows Truncated",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "path",
          "data"
        ],
        "title": "ResultFile"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}